+++ RELEASE_NOTES.rst
@@ -12,6 +12,7 @@ Vagga 0.2.2
 * No need for tilde or null after ``!UbuntuUniverse`` (and probably other cases)
 * Fix permission of ubuntu ``policy-rc.d``, which fixes installing packages
   having a daemon that start on install
+* Configure apt to always use ``--no-install-recommends`` in ubuntu
 
 
 Vagga 0.2.1